+++ src/lib.rs
@@ -78,14 +78,14 @@
 //!
 //! # Usage
 //!
-//! See the [official website](http://alexliesenfeld.github.io/httpmock) a for detailed
+//! See the [official website](http://httpmock.rs) a for detailed
 //! API documentation.
 //!
 //! ## Examples
 //!
 //! You can find examples in the
 //! [`httpmock` test directory](https://github.com/alexliesenfeld/httpmock/blob/master/tests/).
-//! The [official website](http://alexliesenfeld.github.io/httpmock) and [reference docs](https://docs.rs/httpmock/)
+//! The [official website](http://httpmock.rs) and [reference docs](https://docs.rs/httpmock/)
 //! also contain _**a lot**_ of examples.
 //!
 //! ## License